massive update, just commit as it
- Some architecture changes
- Properly rate-limit messages
- Configurator for config file
- Slightly redone modules

The new Main.hs doesn't reflect a change in director of the bot. Rather,
since I molded it into a bot with a specific use case to deal with bans
on #archlinux, a lot of updates where done to the core bot's
architecture. It's easier just to commit this as is then to try to
return to the old Main.hs